fix replace.sh

+++ replace_nvimconfig.sh
@@ -2,6 +2,8 @@
 
 nvim_config_path="$HOME/.config/nvim"
 
-cp $nvim_config_path/init.vim $nvim_config_path/init.vim_old
-cp config_files/init.vim $HOME
+if [[ -f "$nvim_config_path/init.vim" ]]; then
+    cp $nvim_config_path/init.vim $nvim_config_path/init.vim_old
+fi
+cp config_files/init.vim $nvim_config_path
 echo 'config_files/init.vim -> ~/.config/nvim/init.vim'
